vix: split the tree representation layer out of fixture.rs
fixture.rs held two unrelated layers: the offline harness fixture store
(FixtureStore, its entry kinds and read errors, the fixture-tree\0
sentinel, the OriginAdapter impl) and the Tree *representation* layer —
TarMember, UstarParseError, parse_ustar, ResidentTreeError,
tree_from_members, tree_from_resident, canonical_resident_tree, plus
their tests. The latter is value-model machinery consumed by the
untar/tree-glob/exec primitives and the scheduler's canonicalization
path; nothing about it is fixture code, and the origin rail's stage 3
moves FixtureStore out of core while this layer explicitly stays
(origin-rail.md, "What deletes from core").

Moved to a new runtime/tree_resident.rs rather than folded into
tree.rs: tree.rs is the semantic Tree value model (what a Tree *is*);
this layer is transport — how a Tree enters and leaves resident bytes
(ustar archive, carrier, canonical form). Keeping the decode layer a
separate sibling preserves that boundary and keeps tree.rs from
absorbing archive-format parsing.

Pure move: every function body is byte-identical. The module is
glob-re-exported from runtime/mod.rs exactly like fixture was, so all
vix::runtime:: paths (vixen-primitives untar/tree_glob/exec,
primitive.rs, scheduler.rs) are unchanged; the only import touched is
scheduler.rs's explicit `use super::fixture::` line, which now takes
canonical_resident_tree from tree_resident. The moved tests keep using
FixtureStore to load the archive fixture — a test-side convenience
import, not a layer dependency.
